docs(example): update vue-instantsearch example to reflect guide steps
